cote serveur presque achevé avec contraintes bases de donnees

+++ src/main/java/fr/learn/member/MemberController.java
@@ -5,6 +5,7 @@ import java.util.List;
 
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -14,6 +15,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 import fr.learn.dao.Member;
+import fr.learn.dao.Role;
 
 @RestController
 public class MemberController {
@@ -21,12 +23,17 @@ public class MemberController {
 	@Autowired
 	private MemberService memberService;
 	
+	
 	@RequestMapping(value = "/resources/members", method = RequestMethod.POST)
-	public void registerMember(@RequestBody Member member)
+	public boolean registerMember(@RequestBody Member member)
 	{
 		member.setId(null);
 		member.setDateOfRegistration(new Date());
-		memberService.register(member);
+		if(member.getPassword() == null || member.getPassword().length() <6)
+		{
+			return false;
+		}
+		return memberService.register(member);
 	}
 
 	@RequestMapping(value = "/resources/members", method = RequestMethod.GET)
@@ -41,10 +48,24 @@ public class MemberController {
 	
 	
 	@RequestMapping(value = "/resources/members/{idMember}", method = RequestMethod.DELETE)
-	public void deleteMember(@PathVariable("idMember") long idMember)
+	public boolean deleteMember(@PathVariable("idMember") Long idMember)
 	{
-		//TODO only the superUser can drop users
-		memberService.delete(idMember);
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+		Member loggedIn = memberService.getMemberFromAuthentification(auth);
+		if (loggedIn == null)
+			return false;
+		
+		if(loggedIn.getId().equals(idMember))
+			return false;
+		for(Role role : loggedIn.getRoles())
+		{
+			if(role.getName().equals("root"))
+			{
+				memberService.delete(idMember);
+				return true;
+			}
+		}
+		return false;
 	}
 	
 	@RequestMapping(value = "/resources/members/{idMember}", method = RequestMethod.GET)
@@ -57,21 +78,9 @@ public class MemberController {
 	
 	
 	@RequestMapping(value = "/resources/members/{idMember}", method = RequestMethod.PUT)
-	public void updateMember(@PathVariable("idMember") Long idMember, @RequestBody Member member)
+	public boolean updateMember(@PathVariable("idMember") Long idMember, @RequestBody Member member)
 	{
-		//TODO mettre a jour user detailsService apres la mise a jour du model 
-		/*
-		 * 
-		 * Authentication auth = SecurityContextHolder.getContext().setAuthentication();
-		        if(auth.getPrincipal() != null)
-		        {
-		        	
-		        	UserDetails details = (UserDetails)auth.getPrincipal();
-		        	System.out.println(details.getUsername());
-		        	details.setUsername(details.getUsername() +"1");
-		        	
-		        }
-		 */
+		
 		
 		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
 		Member loggedIn = memberService.getMemberFromAuthentification(auth);
@@ -79,8 +88,18 @@ public class MemberController {
 		{
 			member.setId(idMember);
 			member.setPassword(memberService.findOne(idMember).getPassword());
-			memberService.updateMember(member);
+			if(memberService.updateMember(member))
+			{
+				member = memberService.findOne(idMember);
+				Authentication authentication = new UsernamePasswordAuthenticationToken(member.getPseudo(), 
+						member.getPassword(), auth.getAuthorities());
+				SecurityContextHolder.getContext().setAuthentication(authentication);
+				return true;
+			}
+			
 		}
+		
+		return false;
 		
 		
 	}
@@ -96,8 +115,22 @@ public class MemberController {
 		return loggedIn;
 	}
 	
-	
-	
+	@RequestMapping(value = "/resources/members/changePassword", method = RequestMethod.GET)
+	public boolean changeMemberPassword(@RequestBody String password)
+	{
+		if(password == null || password.length() < 6)
+		{
+			return false;
+		}
+		Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+		Member loggedIn = memberService.getMemberFromAuthentification(auth);
+		
+		if(loggedIn == null)
+			return false;
+		
+		memberService.changePasswordForMember(loggedIn, password);
+		return true;
+	}
